requirements.txt
# # Core Libraries
# annotated-types==0.7.0
# blinker==1.9.0
# certifi==2024.8.30
# charset-normalizer==3.4.0
# click==8.1.7
# colorama==0.4.6
# Flask==3.1.0
# idna==3.10
# itsdangerous==2.2.0
# Jinja2==3.1.4
# MarkupSafe==3.0.2
# pillow==11.0.0
# protobuf==5.29.1
# pydantic==2.10.3
# pydantic_core==2.27.1
# pyparsing==3.2.0
# python-dotenv==1.0.1
# requests==2.32.3
# typing_extensions==4.12.2
# urllib3==2.2.3
# Werkzeug==3.1.3
# bs4==0.0.2
# beautifulsoup4==4.12.3
# langchain-openai==0.2.14
# langchain-pinecone==0.2.0
# asyncio==3.4.3

# # Google-related Libraries
# google-ai-generativelanguage==0.6.10
# google-api-core==2.23.0
# google-api-python-client==2.154.0
# google-auth==2.36.0
# google-auth-httplib2==0.2.0
# google-generativeai==0.8.3
# googleapis-common-protos==1.66.0
# grpcio==1.68.1
# grpcio-status==1.68.1
# httplib2==0.22.0
# proto-plus==1.25.0
# uritemplate==4.1.1

# # Other Utilities
# tqdm==4.67.1

# # Optional Development Tools (you might want to remove these in production)
# ipython==8.30.0
# prompt_toolkit==3.0.48


# # Json formatter
# simplejson==3.19.3


### use above if libraries are not working



# Core Libraries
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
asyncio==3.4.3
attrs==24.3.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
protobuf==5.29.1
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3

# HTML Parsing & Web Scraping
bs4==0.0.2
beautifulsoup4==4.12.3
soupsieve==2.6

# AI & Machine Learning
huggingface-hub==0.27.0
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
transformers==4.47.1

# Google-related Libraries
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
proto-plus==1.25.0
uritemplate==4.1.1

# Database & Storage
bson==0.5.10
dnspython==2.7.0
pymongo==4.10.1
fsspec==2024.12.0

# Pinecone Vector Database
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7

# Langchain & LLMs
langchain-core==0.3.28
langchain-google-genai==2.0.7
langchain-openai==0.2.14
langchain-pinecone==0.2.0
langsmith==0.2.4

# Networking & Async HTTP
aiohttp==3.9.5
aiosignal==1.3.2
filelock==3.16.1
httpcore==1.0.7
httpx==0.28.1
requests-toolbelt==1.0.0
yarl==1.18.3

# FastAPI & ASGI Server
uvicorn==0.34.0

# Utility & Miscellaneous
cachetools==5.5.0
executing==2.1.0
filetype==1.2.0
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
propcache==0.2.1
safetensors==0.4.5
six==1.17.0
simplejson==3.19.3
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0

# Progress Bar & Command-line tools
tqdm==4.67.1

# Optional Development Tools (you might want to remove these in production)
ipython==8.30.0
prompt_toolkit==3.0.48
Pygments==2.18.0
matplotlib-inline==0.1.7
traitlets==5.14.3
wcwidth==0.2.13
